Update variantkey with the new header-only version
